backend/requirements.txt
flask
fastapi
transformers
sentence-transformers
faiss-cpu
requests
python-dotenv
numpy<2
datasets